chore: keep only one version of RocketBootstrap
The iOS 15 package's dependencies are broken anyway, it wouldn't be hard
to fix but the only tweak that was known to have issues with CoolStar's
fork is Watusi, which has since been updated, so it's better to just
keep it simple and have one version. This way a working version can also
be automatically installed whenever a tweak depends on RocketBootstrap.